# docs/gophermart/swagger.yaml
basePath: /api/user
definitions:
  domain.UserBalance:
    properties:
      current:
        type: number
      withdrawn:
        type: number
    type: object
  handlers.loginBody:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  handlers.loginResponse:
    properties:
      access_token:
        type: string
    type: object
  handlers.orderForResponse:
    properties:
      accrual:
        type: number
      number:
        type: string
      status:
        type: string
      uploaded_at:
        type: string
    type: object
  handlers.registerBody:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  handlers.registerOrderBody:
    properties:
      order_id:
        type: string
    type: object
  handlers.registerResponse:
    properties:
      access_token:
        type: string
    type: object
  handlers.userWithdrawalForResponse:
    properties:
      order:
        type: string
      processed_at:
        type: string
      sum:
        type: number
    type: object
  handlers.withdrawBalanceBody:
    properties:
      order:
        type: string
      sum:
        type: number
    type: object
  httputils.HTTPError:
    properties:
      error:
        type: string
    type: object
info:
  contact: {}
  description: Gophermart Loyalty Service responsible for saving user orders, saving
    user balance and withdraw balance
  title: Gophermart Loyalty Service
  version: "1.0"
paths:
  /balance:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UserBalance'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.HTTPError'
      security:
      - BearerAuth: []
      summary: Get user balance
      tags:
      - balance
  /balance/withdraw:
    post:
      consumes:
      - application/json
      parameters:
      - description: Withdraw from balance
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/handlers.withdrawBalanceBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputils.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputils.HTTPError'
        "402":
          description: Payment Required
          schema:
            $ref: '#/definitions/httputils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.HTTPError'
      security:
      - BearerAuth: []
      summary: Withdraw balance from account
      tags:
      - balance
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login to account
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/handlers.loginBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.loginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputils.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.HTTPError'
      summary: Login to account by credentials
      tags:
      - auth
  /orders:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.orderForResponse'
            type: array
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.HTTPError'
      security:
      - BearerAuth: []
      summary: Get user registered orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      parameters:
      - description: Register order in system
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/handlers.registerOrderBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputils.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputils.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httputils.HTTPError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/httputils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.HTTPError'
      security:
      - BearerAuth: []
      summary: Register order in loyalty system
      tags:
      - orders
  /register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Register new user
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/handlers.registerBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.registerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputils.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httputils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.HTTPError'
      summary: Register new user
      tags:
      - auth
  /withdrawals:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.userWithdrawalForResponse'
            type: array
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputils.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputils.HTTPError'
      security:
      - BearerAuth: []
      summary: Get user withdrawals history
      tags:
      - balance
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"